backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic==2.5.0
sentence-transformers==2.2.2
anthropic==0.7.8
python-dotenv==1.0.0
click==8.1.7
numpy==1.24.3
pandas==2.1.4
sqlalchemy==2.0.23
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1